Unity Build WIP - CommonIncludes OK

+++ Source/Effect/effects/time/TimedEffect.h
@@ -28,7 +28,7 @@ public:
 
 
 	double timeAtLastUpdate;
-	double curTime;
+	float curTime;
 
 	virtual void updateEnabled() override;
